Rename providers->media-url-parser
